// ui/src/generated/types/conflict.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */
import type { ConflictDate } from './conflictDate'
import type { ConflictOwner } from './conflictOwner'

export interface Conflict {
  /** Timeslot in conflict */
  date?: ConflictDate
  /** Devices in conflict */
  devices?: string[]
  /** Name of the group in conflict */
  group?: string
  /** Owner of the group in conflict */
  owner?: ConflictOwner
}